requirements.txt
colorama >= 0.4.5
requests >= 2.28.1
cerberus >= 1.3.4
pyyaml >= 6.0
pygame >= 2.1.2
